tidying  the project structure
