+++ src/app.ts
@@ -1,11 +1,11 @@
 import express from 'express';
 import { dbConnection } from '@database/index';
-import ProductRouts from '@routes/product';
+import ProductRoutes from '@routes/product';
 
 const app = express();
 
 dbConnection.sync({ alter: false });
 
-app.use('/product', ProductRouts);
+app.use('/product', ProductRoutes);
 
 export default app;
